Adding crop photo

+++ src/main/java/ru/todo100/activer/service/ServiceAbstract.java
@@ -25,7 +25,7 @@ abstract public class ServiceAbstract {
 		return session;
 	}
 	
-	final public void delete(Long id) {
+	final public void delete(Integer id) {
 		Object object = getSession().get(getItemClass(), id);
 		if (object!=null) {
 			getSession().delete(object);
